+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 + Support inferring types of a constructor call
 + Merge `normal.rs` into `read_back.rs`
 + Subtyping: `sum { A }` is a subtype of `sum { A | B }`
++ Add tutorial in `lib.rs`
 
 # 0.2.2